requirements/dev.txt
Django==1.7.1
GitPython==0.3.2.RC1
argparse==1.2.1
async==0.6.1
coverage==3.7.1
django-bootstrap3==4.11.0
django-contrib-comments==1.5
django-epiceditor==0.2.3
django-gravatar2==1.1.4
django-nose==1.2
django-taggit==0.12.2
django-timezone-field==1.1
gitdb==0.5.4
ipython==2.3.0
model-mommy==1.2.2
nose==1.3.4
pytz==2014.10
requests==2.5.0
six==1.7.3
smmap==0.8.2
wsgiref==0.1.2
